refactor: WriterResponse 위치 변경
